Colours on pics still sus
I got a Notebook to work where it saves the model and only trains it if the model hasn't been save.
Thus the Notebook can always be run by just pressing "Run All" instead of a selection of code cells.
Though, the colours after the adversarial attack are still might sussy.