Both ResNet and VGG should be ready for GBAR
Updated torchAttacks notebooks #2, -#3 and -#4 with versatile checkpoints. As accuracy and loss is saved in the dictionary alongside the model, saving figures is now less of a priority.
Also added msg( ) function, for nicer prints to terminal.
Moved some outdated files to the oldFiles directory as I believe they contain a hyperlink to a good reference for saving models (which we now do differently).