Mikkel's linux and GBAR magic
Probably should have pushed Saturday. I updated the repo, with a notebook that now has a validation set (acquired from a 80-20 split of the train set). Furthermore, additional updates include: Mikkel's GBAR magic, to some extend at least, and two different tests, to see whether I use the right Virtual Environment on GBAR. Which I now do after a lot struggle.

The queue-job.sh file is a bash script, that is a pseudo template for the script that is auto-pushed to the queue-thingy on GBAR.