That's a lot of DAMAGE!
Wanted to wrote that it was a lot of changes and then remembered that quote, which is utterly useless for this commit.
Think I got this shizzle to work. Added some scripts that clearly doesn't work - mainly just to have them versioned control. In other words:
"Felt cute, might delete later".